Fill README.md, change seaLevel type to int, modify band map renderer, fix freeing tectonic vectors

+++ src/config.h
@@ -12,7 +12,7 @@ typedef struct {
     float seaPlateHeight;
     float landPlateHeight;
     float tectonicImpactFactor;
-    float seaLevel;
+    int seaLevel;
     int gaussianRange;
     float gaussianDiminishingFactor;
     int tectonicCount;
